fixing JS load bug
