+++ app/assets/stylesheets/common.scss
@@ -12,11 +12,9 @@
 }
 
 .label_photo {
-  width: 100px;
-  height: 100px;
-  background-color: red;
+  width: 150px;
+  height: 150px;
   display: block;
   background-position: center;
   background-size: cover;
-  border-radius: 50%;
 }
